feat: simplifying the approach
