+++ bundles/com.espressif.idf.core/META-INF/MANIFEST.MF
@@ -38,14 +38,11 @@ Export-Package: com.espressif.idf.core,
  com.espressif.idf.core.toolchain,
  com.espressif.idf.core.tools,
  com.espressif.idf.core.tools.eimjson,
- com.espressif.idf.core.tools.eimjson.adapter,
  com.espressif.idf.core.tools.eimjson.model,
  com.espressif.idf.core.tools.eimjson.presentation,
- com.espressif.idf.core.tools.eimjson.schema.v3,
  com.espressif.idf.core.tools.exceptions,
  com.espressif.idf.core.tools.launch,
  com.espressif.idf.core.tools.util,
- com.espressif.idf.core.tools.vo,
  com.espressif.idf.core.tools.watcher,
  com.espressif.idf.core.util,
  com.espressif.idf.core.variable,
